requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
cryptography==36.0.2
fpdf2==2.8.3
Jinja2==3.1.2
gunicorn==23.0.0
psycopg2-binary # Adicionado para suporte PostgreSQL

